Dragon no longer moves one more time after victory requirements are met
Due to the way the end-of-turn logic was implemented, the Dragon was moving even if a player had enough points to win the game. In some situations, the Dragon would move onto the same space as the "winning" Player, causing the Player to drop their Poo and return to their starting area.